+++ federatedscope/core/data/utils.py
@@ -4,6 +4,7 @@ import logging
 import os
 import re
 from collections import defaultdict
+from os import path as osp
 
 import numpy as np
 from random import shuffle
@@ -615,3 +616,35 @@ def merge_data(all_data, merged_max_data_id=None, specified_dataset_name=None):
             " 1): {data_id: {train: {x:ndarray, y:ndarray}} }"
             " 2): {data_id: {train: DataLoader }")
     return merged_data
+
+
+def save_local_data(dir_path,
+                    train_data=None,
+                    train_targets=None,
+                    test_data=None,
+                    test_targets=None,
+                    val_data=None,
+                    val_targets=None):
+    r"""
+    https://github.com/omarfoq/FedEM/blob/main/data/femnist/generate_data.py
+
+    save (`train_data`, `train_targets`) in {dir_path}/train.pt,
+    (`val_data`, `val_targets`) in {dir_path}/val.pt
+    and (`test_data`, `test_targets`) in {dir_path}/test.pt
+    :param dir_path:
+    :param train_data:
+    :param train_targets:
+    :param test_data:
+    :param test_targets:
+    :param val_data:
+    :param val_targets
+    """
+    import torch
+    if (train_data is not None) and (train_targets is not None):
+        torch.save((train_data, train_targets), osp.join(dir_path, "train.pt"))
+
+    if (test_data is not None) and (test_targets is not None):
+        torch.save((test_data, test_targets), osp.join(dir_path, "test.pt"))
+
+    if (val_data is not None) and (val_targets is not None):
+        torch.save((val_data, val_targets), osp.join(dir_path, "val.pt"))
