+++ federatedscope/core/trainers/__init__.py
@@ -1,6 +1,7 @@
 from federatedscope.core.trainers.base_trainer import BaseTrainer
 from federatedscope.core.trainers.trainer import Trainer
 from federatedscope.core.trainers.torch_trainer import GeneralTorchTrainer
+from federatedscope.core.trainers.tf_trainer import GeneralTFTrainer
 from federatedscope.core.trainers.trainer_multi_model import \
     GeneralMultiModelTrainer
 from federatedscope.core.trainers.trainer_pFedMe import wrap_pFedMeTrainer
@@ -15,5 +16,5 @@ __all__ = [
     'Trainer', 'Context', 'GeneralTorchTrainer', 'GeneralMultiModelTrainer',
     'wrap_pFedMeTrainer', 'wrap_DittoTrainer', 'FedEMTrainer',
     'wrap_fedprox_trainer', 'wrap_nbafl_trainer', 'wrap_nbafl_server',
-    'BaseTrainer'
+    'BaseTrainer', 'GeneralTFTrainer'
 ]
